Add safe/non-desctructive mode to manage.py (#32)
* Add a non-desctructive mode to keep manually added labels
* Run manage.py in safe/non-desctructive mode
* Make safe mode false by default. We invoke it with --safe though.
* Fix how manage.py is invoked and standardize environment
* Rename safe option to keep_labels

+++ manage.py
@@ -13,6 +13,7 @@ gh = github.Github(login_or_token=API_TOKEN)
 
 parser = ArgumentParser()
 parser.add_argument("-d", "--dry", default=False, help="Dry run - if true does not change github items")
+parser.add_argument("-kl", "--keep_labels", default=False, help="If true does not remove manually added, non-defined labels")
 args = parser.parse_args()
 
 with open("config.yaml") as fp:
@@ -60,7 +61,7 @@ for gh_repository in gh.get_organization(ORGANIZATION).get_repos(type='public'):
 
     for label in labels.keys():
         logging.info(f"{gh_repository.name} - {label} should not exist")
-        if args.dry is False:
+        if args.dry is False and args.keep_labels is False:
             gh_label = labels[label]
             gh_label.delete()
             logging.info(f"{gh_repository.name} - {label} removed")
